Modularizing Collection resource
`Service.js` manages code interacting with the database
`Schema.js` ensures standard data delivery
Added Joi dependency for `Schema.js`
