+++ package.json
@@ -58,6 +58,7 @@
     "good-squeeze": "^5.0.1",
     "hapi": "^16.1.0",
     "inert": "^4.1.0",
+    "joi": "^10.4.1",
     "nunjucks-hapi": "^2.1.0",
     "pg": "^6.1.2",
     "pg-hstore": "^2.3.2",
